+++ app/src/main/java/com/example/nermingraca/bitbayapp/ShowProducts.java
@@ -65,7 +65,7 @@ public class ShowProducts extends ActionBarActivity {
 
             @Override
             public void onTextChanged(CharSequence s, int start, int before, int count) {
-                ( (ArrayAdapter<Product>)mProductList.getAdapter()).getFilter().filter(s);
+                ( (CustomListAdapter)mProductList.getAdapter()).getFilter().filter(s);
             }
 
             @Override
@@ -73,78 +73,6 @@ public class ShowProducts extends ActionBarActivity {
 
             }
         });
-    }
-
-    private class ProductsAdapter extends ArrayAdapter<Product> {
-
-        private ArrayList<Product> mListToShow;
-        private Filter mFilter;
-
-        private ProductsAdapter(Context context, int resource, ArrayList<Product> origin) {
-            super(context, resource);
-            mListToShow = origin;
-        }
-
-        @Override
-        public Filter getFilter(){
-            if(mFilter == null) {
-                mFilter = new PostsFilter();
-            }
-            return mFilter;
-        }
-
-        @Override
-        public int getCount() {
-            if (mListToShow != null) {
-                return mListToShow.size();
-            } else {
-                return 0;
-            }
-
-        }
-
-        @Override
-        public Product getItem(int position){
-            return mListToShow.get(position);
-        }
-
-        private class PostsFilter extends Filter {
-
-            @Override
-            protected FilterResults performFiltering(CharSequence constraint) {
-
-                FilterResults results = new FilterResults();
-
-                if(constraint == null || constraint.length() == 0) {
-                    List<Product> origin = ProductFeed.getInstance().getFeed();
-                    results.values = origin;
-                    results.count = origin.size();
-                } else {
-                    String searchString = constraint.toString().toLowerCase();
-                    ArrayList<Product> filteredList = new ArrayList<Product>();
-                    for(int i = 0; i < mListToShow.size(); i++) {
-                        Product p = mListToShow.get(i);
-                        String postTitle = p.getmName().toLowerCase();
-
-                        if(postTitle.contains(searchString)) {
-                            filteredList.add(p);
-                        }
-                        results.values = filteredList;
-                        results.count = filteredList.size();
-                    }
-                }
-
-                return results;
-            }
-
-            @Override
-            protected void publishResults(CharSequence constraint, FilterResults results) {
-                if(results.values != null) {
-                    mListToShow = (ArrayList<Product>)results.values;
-                    notifyDataSetChanged();
-                }
-            }
-        }
     }
